Move diskLimiter up
